pyproject.toml
[project]
name = "langchain-tutorials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "devtools>=0.12.2",
    "faiss-gpu",
    "gradio>=6.10.0",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langsmith[otel]>=0.7.22",
    "openai>=2.30.0",
    "openevals>=0.1.3",
    "opentelemetry-exporter-otlp>=1.40.0",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.2.2",
]
